feat: list active proposals and allow to veto them
