+++ package.json
@@ -26,6 +26,7 @@
     "assert": "2.1.0",
     "buffer": "6.0.3",
     "crypto-browserify": "3.12.0",
+    "dayjs": "^1.11.10",
     "ethers": "5.7.1",
     "graphql": "^16.8.1",
     "https-browserify": "1.0.0",
